--- Proj1/C/Tree.c
@@ -1,3 +1,0 @@
-#include "Tree.h"
-
-void addNode(Tree, char binary )
