--- Proj1/C/Tree.h
@@ -1,12 +1,0 @@
-#IFNDEF Tree
-#DEF TREE
-typedef struct{
-    int value;
-    node * left;
-    node * right;
-}node;
-typedef struct{
-    node * root
-}Tree;
-
-#ENDIF
