--- Proj1/datasetA.txt
@@ -1,7 +1,0 @@
-0 2
-00 3
-10 4
-11 5
-010 5
-1001 4
-
