--- Proj1/datasetB.txt
@@ -1,7 +1,0 @@
-0 2
-00 3
-10 4
-010 5
-1001 6
-1101 7
-
