+++ projeto-1/C/Tree.c
@@ -1,0 +1,3 @@
+#include "Tree.h"
+
+void addNode(Tree, char binary )
