+++ projeto-1/C/Tree.h
@@ -1,0 +1,12 @@
+#IFNDEF Tree
+#DEF TREE
+typedef struct{
+    int value;
+    node * left;
+    node * right;
+}node;
+typedef struct{
+    node * root
+}Tree;
+
+#ENDIF
